requirements.txt
python-vlc
flask
RPi.GPIO
